+++ cookbooks/app-react-bookmark-advanced/src/App.tsx
@@ -1,11 +1,8 @@
-import { StrictMode } from 'react';
 import Router from './Router';
 
 export const App = () => {
     return (
-        <StrictMode>
-            <Router />
-        </StrictMode>
+        <Router />
     );
 };