+++ packages/modules/context/src/utils/enable-context.ts
@@ -3,10 +3,10 @@ import type {
     AnyModule,
     ModuleInitializerArgs,
 } from '@equinor/fusion-framework-module';
-import type { IContextModuleConfigurator } from './configurator';
-import type { ContextConfigBuilder } from './ContextConfigBuilder';
+import type { IContextModuleConfigurator } from '../configurator';
+import type { ContextConfigBuilder } from '../ContextConfigBuilder';
 
-import { module } from './module';
+import { module } from '../module';
 
 /**
  * Method for enabling the Service module