+++ packages/react/legacy-interopt/src/static.ts
@@ -1,1 +1,2 @@
-export const globalEquinorFusionContextKey = '74b1613f-f22a-451b-a5c3-1c9391e91e68';
+export const GLOBAL_FUSION_CONTEXT_KEY = '74b1613f-f22a-451b-a5c3-1c9391e91e68';
+export const LOCAL_STORAGE_CURRENT_CONTEXT_KEY = 'FUSION_CURRENT_CONTEXT';