+++ packages/react/modules/bookmark/src/useBookmark.ts
@@ -3,11 +3,17 @@ import {
     PatchBookmark,
     BookmarkModule,
     UpdateBookmarkOptions,
+    IBookmarkModuleProvider,
 } from '@equinor/fusion-framework-module-bookmark';
 import { useFramework } from '@equinor/fusion-framework-react';
 import { useObservableState } from '@equinor/fusion-observable/react';
 import { useCallback, useMemo } from 'react';
 import { Bookmarks, CreateBookMarkFn } from './types';
+
+import { EMPTY } from 'rxjs';
+
+const useBookmarkProvider = (): IBookmarkModuleProvider | undefined =>
+    useFramework<[BookmarkModule]>().modules.bookmark;
 
 /**
  *  For application development the useCurrentBookmark should be sufficient enough
@@ -26,17 +32,17 @@ import { Bookmarks, CreateBookMarkFn } from './types';
  * @return {*}  {Bookmarks<TData>} the full api fro handling bookmarks
  */
 export const useBookmark = <TData>(): Bookmarks<TData> => {
-    const bookmarkProvider = useFramework<[BookmarkModule]>().modules.bookmark;
+    const bookmarkProvider = useBookmarkProvider();
 
     const currentBookmark = useObservableState(
-        useMemo(() => bookmarkProvider.currentBookmark$, [bookmarkProvider]),
+        useMemo(() => bookmarkProvider?.currentBookmark$ ?? EMPTY, [bookmarkProvider]),
         {
-            initial: bookmarkProvider.currentBookmark,
+            initial: bookmarkProvider?.currentBookmark,
         }
-    ).value as Bookmark<TData>;
+    ).value as Bookmark<TData> | null | undefined;
 
     const bookmarks = useObservableState(
-        useMemo(() => bookmarkProvider.bookmarks$, [bookmarkProvider]),
+        useMemo(() => bookmarkProvider?.bookmarks$ ?? EMPTY, [bookmarkProvider]),
         {
             initial: [],
         }
@@ -48,7 +54,7 @@ export const useBookmark = <TData>(): Bookmarks<TData> => {
                 return () => {
                     // Noting to remove
                 };
-            return bookmarkProvider.addStateCreator(() => {
+            return bookmarkProvider?.addStateCreator(() => {
                 return createBookmarkState();
             });
         },
@@ -60,8 +66,8 @@ export const useBookmark = <TData>(): Bookmarks<TData> => {
             name: string;
             description: string;
             isShared: boolean;
-        }): Promise<Bookmark<T>> => {
-            return await bookmarkProvider.createBookmark<T>(args);
+        }): Promise<Bookmark<T> | undefined> => {
+            return bookmarkProvider && (await bookmarkProvider.createBookmark<T>(args));
         },
         [bookmarkProvider]
     );
@@ -70,48 +76,56 @@ export const useBookmark = <TData>(): Bookmarks<TData> => {
             bookmark: PatchBookmark<T>,
             options?: UpdateBookmarkOptions
         ): Promise<Bookmark<T> | undefined> => {
-            return await bookmarkProvider.updateBookmarkAsync<T>(bookmark, options);
+            return (
+                bookmarkProvider &&
+                (await bookmarkProvider.updateBookmarkAsync<T>(bookmark, options))
+            );
         },
         [bookmarkProvider]
     );
     const deleteBookmarkById = useCallback(
-        async (bookmarkId: string): Promise<string> => {
-            return await bookmarkProvider.deleteBookmarkByIdAsync(bookmarkId);
+        async (bookmarkId: string): Promise<string | undefined> => {
+            return bookmarkProvider && (await bookmarkProvider.deleteBookmarkByIdAsync(bookmarkId));
         },
         [bookmarkProvider]
     );
     const getAllBookmarks = useCallback(async (): Promise<Array<Bookmark>> => {
-        return await bookmarkProvider.getAllBookmarksAsync();
+        return bookmarkProvider ? await bookmarkProvider.getAllBookmarksAsync() : [];
     }, [bookmarkProvider]);
 
     const addBookmarkFavorite = useCallback(
         async (bookmarkId: string): Promise<void> => {
-            return await bookmarkProvider.addBookmarkFavoriteAsync(bookmarkId);
+            return (
+                bookmarkProvider && (await bookmarkProvider.addBookmarkFavoriteAsync(bookmarkId))
+            );
         },
         [bookmarkProvider]
     );
     const removeBookmarkFavorite = useCallback(
         async (bookmarkId: string): Promise<void> => {
-            return await bookmarkProvider.removeBookmarkFavoriteAsync(bookmarkId);
+            return (
+                bookmarkProvider && (await bookmarkProvider.removeBookmarkFavoriteAsync(bookmarkId))
+            );
         },
         [bookmarkProvider]
     );
 
     const setCurrentBookmark = useCallback(
         <TData>(IdOrItem: string | Bookmark<TData>): void => {
-            bookmarkProvider.setCurrentBookmark(IdOrItem);
+            bookmarkProvider && bookmarkProvider.setCurrentBookmark(IdOrItem);
         },
         [bookmarkProvider]
     );
 
     const getBookmarkById = useCallback(
-        <TData>(id: string): Promise<Bookmark<TData>> =>
-            bookmarkProvider.getBookmarkById<TData>(id),
+        <TData>(id: string): Promise<Bookmark<TData>> | undefined =>
+            bookmarkProvider && bookmarkProvider.getBookmarkById<TData>(id),
         [bookmarkProvider]
     );
 
     const getCurrentAppKey = useCallback(() => {
         return (
+            bookmarkProvider &&
             bookmarkProvider.config.getCurrentAppIdentification &&
             bookmarkProvider.config.getCurrentAppIdentification()
         );