+++ packages/utils/observable/src/types/index.ts
@@ -1,3 +1,5 @@
+import { Observable } from 'rxjs';
+
 export * from './actions';
 export * from './flow';
 export * from './observable';
@@ -5,3 +7,5 @@ export * from './reducers';
 
 /** @deprecated use {@link Flow}  */
 export { Flow as Epic } from './flow';
+
+export type ObservableType<T> = T extends Observable<infer U> ? U : never;