+++ packages/modules/bookmark/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.6](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-module-bookmark@1.0.5...@equinor/fusion-framework-module-bookmark@1.0.6) (2023-05-23)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-bookmark
+
+## 1.0.5 (2023-05-22)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-bookmark
+
+## [1.0.4](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-module-bookmark@1.0.3...@equinor/fusion-framework-module-bookmark@1.0.4) (2023-05-05)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-bookmark
+
 ## 1.0.3 (2023-04-24)
 
 ### Bug Fixes