+++ packages/modules/widget/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.3](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-module-widget@1.0.2...@equinor/fusion-framework-module-widget@1.0.3) (2023-05-23)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-widget
+
+## 1.0.2 (2023-05-22)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-widget
+
+## 1.0.1 (2023-05-05)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-widget
+
 ## 1.0.0 (2023-04-16)
 
 **Note:** Version bump only for package @equinor/fusion-framework-module-widget