+++ packages/react/components/bookmark/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.11](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-components-bookmark@0.1.10...@equinor/fusion-framework-react-components-bookmark@0.1.11) (2023-05-24)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-components-bookmark
+
+## [0.1.10](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-components-bookmark@0.1.9...@equinor/fusion-framework-react-components-bookmark@0.1.10) (2023-05-23)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-components-bookmark
+
+## [0.1.9](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-components-bookmark@0.1.8...@equinor/fusion-framework-react-components-bookmark@0.1.9) (2023-05-22)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-components-bookmark
+
+## [0.1.8](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-components-bookmark@0.1.7...@equinor/fusion-framework-react-components-bookmark@0.1.8) (2023-05-22)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-components-bookmark
+
+## 0.1.7 (2023-05-22)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-components-bookmark
+
+## [0.1.6](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-components-bookmark@0.1.5...@equinor/fusion-framework-react-components-bookmark@0.1.6) (2023-05-12)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-components-bookmark
+
+## [0.1.5](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-components-bookmark@0.1.4...@equinor/fusion-framework-react-components-bookmark@0.1.5) (2023-05-11)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-components-bookmark
+
+## [0.1.4](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-components-bookmark@0.1.3...@equinor/fusion-framework-react-components-bookmark@0.1.4) (2023-05-11)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-components-bookmark
+
+## 0.1.3 (2023-05-10)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-components-bookmark
+
+## [0.1.2](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-components-bookmark@0.1.1...@equinor/fusion-framework-react-components-bookmark@0.1.2) (2023-05-08)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-components-bookmark
+
+## [0.1.1](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-react-components-bookmark@0.1.0...@equinor/fusion-framework-react-components-bookmark@0.1.1) (2023-05-05)
+
+**Note:** Version bump only for package @equinor/fusion-framework-react-components-bookmark
+
 ## 0.1.0 (2023-04-24)
 
 ### Features