+++ packages/widget/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.13](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-widget@1.0.12...@equinor/fusion-framework-widget@1.0.13) (2023-05-23)
+
+**Note:** Version bump only for package @equinor/fusion-framework-widget
+
+## 1.0.12 (2023-05-22)
+
+**Note:** Version bump only for package @equinor/fusion-framework-widget
+
+## [1.0.11](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-widget@1.0.10...@equinor/fusion-framework-widget@1.0.11) (2023-05-12)
+
+**Note:** Version bump only for package @equinor/fusion-framework-widget
+
+## [1.0.10](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-widget@1.0.9...@equinor/fusion-framework-widget@1.0.10) (2023-05-11)
+
+**Note:** Version bump only for package @equinor/fusion-framework-widget
+
+## [1.0.9](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-widget@1.0.8...@equinor/fusion-framework-widget@1.0.9) (2023-05-11)
+
+**Note:** Version bump only for package @equinor/fusion-framework-widget
+
+## 1.0.8 (2023-05-10)
+
+**Note:** Version bump only for package @equinor/fusion-framework-widget
+
+## [1.0.7](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-widget@1.0.6...@equinor/fusion-framework-widget@1.0.7) (2023-05-08)
+
+**Note:** Version bump only for package @equinor/fusion-framework-widget
+
+## [1.0.6](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-widget@1.0.5...@equinor/fusion-framework-widget@1.0.6) (2023-05-05)
+
+**Note:** Version bump only for package @equinor/fusion-framework-widget
+
 ## 1.0.5 (2023-04-24)
 
 **Note:** Version bump only for package @equinor/fusion-framework-widget